Un-nest input bar component
