--- front-end/src/components/input-bar/InputBar.tsx
@@ -1,22 +1,0 @@
-import { InputButton } from "./InputButton.js";
-
-import styles from "./InputBar.module.css";
-import pencil from "../../assets/pencil.svg";
-import paperAirplane from "../../assets/paper-airplane.svg";
-
-export function InputBar () {
-	function openKeyboard () {
-		alert("You should probably change this!");
-	}
-	function send () {
-		alert("You should probably change this!");
-	}
-
-	return (
-		<div className={styles.inputBar}>
-			<InputButton icon={pencil} clicked={openKeyboard} />
-			<input type="text" placeholder="Type anything here..." className={styles.textBox}/>
-			<InputButton icon={paperAirplane} clicked={send} />
-		</div>
-	);
-}
