--- front-end/src/components/input-bar/InputButton.tsx
@@ -1,14 +1,0 @@
-import styles from "./InputBar.module.css";
-
-interface Props {
-    icon: string;
-    clicked?: () => void;
-}
-
-export function InputButton ({ icon, clicked }: Props) {
-	return (
-		<div className={styles.button} onClick={clicked}>
-			<img src={icon} alt="icon" className={styles.buttonIcon} />
-		</div>
-	);
-}
